Checking if any openshift_*_storage_kind variables are set to dynamic without enabling dynamic provisioning
